refactor(ocr): add lazy-loading for ocr and move OCR helpers into FontOCR class
