+++ src/novel_downloader/core/__init__.py
@@ -7,12 +7,11 @@ This package serves as the core layer of the novel_downloader system.
 
 It provides factory methods for constructing key components required for
 downloading and processing online novel content, including:
-
-- Downloader: Handles the full download lifecycle of a book or a batch of books.
-- Parser: Extracts structured data from HTML or SSR content.
-- Fetcher: Sends HTTP requests and manages sessions, including login if required.
-- Exporter: Responsible for exporting downloaded data into various output formats.
-- search: Provides unified search functionality across supported novel sites.
+  * Downloader: Handles the full download lifecycle of a book or a batch of books.
+  * Parser: Extracts structured data from HTML or SSR content.
+  * Fetcher: Sends HTTP requests and manages sessions, including login if required.
+  * Exporter: Responsible for exporting downloaded data into various output formats.
+  * search: Provides unified search functionality across supported novel sites.
 """
 
 __all__ = [
