+++ src/novel_downloader/core/parsers/base.py
@@ -43,6 +43,7 @@ class BaseParser(ParserProtocol, abc.ABC):
         self._config = config
         self._book_id: str | None = None
 
+        self._save_font_debug = config.save_font_debug
         self._decode_font: bool = config.decode_font
         self._use_truncation = config.use_truncation
         self._base_cache_dir = Path(config.cache_dir)
