+++ src/novel_downloader/locales/zh.json
@@ -39,9 +39,9 @@
   "login_description": "说明",
   "login_hint": "提示",
   "login_use_config": "使用配置中的默认值",
-  "login_enter_password": "请输入密码: ",
-  "login_enter_cookie": "请输入 Cookie: ",
-  "login_enter_value": "请输入值: ",
+  "login_enter_password": "请输入密码",
+  "login_enter_cookie": "请输入 Cookie",
+  "login_enter_value": "请输入值",
   "login_required_field": "该字段是必填项, 请重新输入",
 
   "clean_failed": "删除失败: {path}",
